Pull certificates directly from certifi's source
Should simplify the build process a bit as it no longer relies on
Python. Also no longer pulls in OpenSSL. So we shouldn't need to worry
about file conflicts. It also simplifies the upgrade process as it no
longer relies on an upgrade to certifi to occur first. Finally we can
package the license file now, which should ease some concerns.